+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-09-01 18:34:12
-// Commit: test(test): implement widget integration tests for test
+// Updated at: 2026-09-01 19:40:47
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '4120';
-  final String lastMessage = 'test(test): implement widget integration tests for test';
-  final DateTime timestamp = DateTime.parse('2026-09-01T18:34:12+05:00');
+  final String commitRef = '4123';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-09-01T19:40:47+05:00');
 }